Move instrument name validation into InstrumentBuilder
Review feedback: the validator was a small file whose logic is not used
anywhere else, so it lives better alongside its only two call sites.

InstrumentNameValidation.isValid / .maximumLength become
InstrumentBuilder.isValidName / .maximumNameLength, and warnIfInvalid
becomes an instance method that reads the builder's own instrumentName.
No behaviour change.
